refactor: reorganize variable input files, moved to helpers
