+++ frontend/src/components/input.tsx
@@ -9,7 +9,7 @@ import {
   createVariableKeyDownHandler,
   createVariableKeyUpHandler,
   createVariableMouseUpHandler,
-} from "./variable-input"
+} from "../helpers/variable-input"
 
 const INPUT_BASE_CLASS =
   "flex h-10 w-full rounded-full border border-input bg-transparent px-3 py-2 text-base ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium file:text-foreground placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 md:text-sm pointer-events-auto relative z-10"
