+++ frontend/src/helpers/variable-input/VariableOverlayContent.tsx
