+++ frontend/src/helpers/variable-input/index.ts
