+++ frontend/src/helpers/variable-input/templateVariableCaret.ts
