+++ frontend/src/helpers/variable-input/templateVariableConstants.ts
