+++ frontend/src/helpers/variable-input/templateVariableHandlers.ts
