+++ frontend/src/helpers/variable-input/templateVariableHighlight.ts
