+++ frontend/src/helpers/variable-input/templateVariableRanges.ts
